+++ src/Resources/Policy.Autorest/internal/New-AzPolicyDefinitionVersion.ps1
@@ -75,7 +75,8 @@ param(
 
     [Parameter()]
     [Microsoft.Azure.PowerShell.Cmdlets.Policy.Category('Body')]
-    [Microsoft.Azure.PowerShell.Cmdlets.Policy.Models.IAny]
+    [Microsoft.Azure.PowerShell.Cmdlets.Policy.Runtime.Info(PossibleTypes=([Microsoft.Azure.PowerShell.Cmdlets.Policy.Models.IPolicyDefinitionVersionPropertiesPolicyRule]))]
+    [System.Collections.Hashtable]
     # The policy rule.
     ${PolicyRule},