+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Upgraded Microsoft.Azure.Storage.DataMovement to 2.0.5
+
+## Version 7.1.0
 * Fixed the issue that Azure.Core.AccessToken is used before assigned.
 * Supported TLS1_3 when creating and updating a storage account 
     - `New-AzStorageAccount`